Phase 6: dev-harness CLI subcommands + supporting changes
Add the wwcli dev-harness subcommands and the runtime/build changes that
back them:

  crosscheck  - clip cross-check vs a reference collision_map.bin
  walltest    - wall-shape reflection/rotation invariants
  scripted    - named path-correctness categories (walk/plane/teleport/gate)
  bench       - query benchmarks
  path        - JSON path export for the Leaflet viewer
  doors       - same-plane Transport ("door") crossing traversal, with a
                doorprobe helper that dumps origin/dest collision geometry

The doors sweep confirms ~445/451 area-splitting crossings route through
the object correctly; the handful that don't are shape-10 stepping-stone /
interactive-scenery objects whose origin sits on a blocked tile (part of
the builder's reported unresolved-origin set; ADR 0009 "dropped and logged
for dataset review"), not wall doors.

+++ src/runtime/WorldView.h
@@ -61,8 +61,12 @@ namespace ww::runtime
         const std::vector<int32_t> &gridIds(int squareX, int squareY, int plane);
 
         const format::ArtifactReader *artifact;
-        std::unordered_map<uint32_t, std::vector<uint32_t>> clipCache;
-        std::unordered_map<uint32_t, std::vector<int32_t>> gridCache;
+        // 64-bit non-overlapping keys (squareY << 32 | squareX, and
+        // squareY << 40 | squareX << 8 | plane) so squareX, squareY, and plane
+        // can never alias under any cache size — RS3 has at most ~128 squares
+        // per axis but the room is free.
+        std::unordered_map<uint64_t, std::vector<uint32_t>> clipCache;
+        std::unordered_map<uint64_t, std::vector<int32_t>> gridCache;
     };
 }
